+++ source/dea-app/package.json
@@ -53,6 +53,7 @@
     "@types/jest": "^27.4.1",
     "@types/lodash": "^4.14.181",
     "@types/node": "^14",
+    "@types/supertest": "^2.0.12",
     "@types/uuid": "^8.3.4",
     "concurrently": "^7.0.0",
     "depcheck": "^1.4.3",
@@ -67,7 +68,9 @@
     "npm-package-json-lint": "^6.3.0",
     "npm-package-json-lint-config-default": "^5.0.0",
     "sort-package-json": "^1.57.0",
+    "supertest": "^6.2.4",
     "ts-jest": "^27.1.3",
+    "ts-mockito": "~2.6.1",
     "typescript": "^4.5.2"
   }
 }
