+++ source/dea-backend/src/backendAPILambda.ts
@@ -3,7 +3,7 @@
  *  SPDX-License-Identifier: Apache-2.0
  */
 
+import { getBackendApiApp } from '@aws/dea-app';
 import serverlessExpress from '@vendia/serverless-express';
-import { getBackendApiApp } from './backendAPI';
 
 exports.handler = serverlessExpress({ app: getBackendApiApp() });
